requirements.txt
SQLAlchemy>=2.0,<3
python-dotenv>=1.0,<2

# Optional DB drivers (pure-Python, work well on Windows)
PyMySQL>=1.1,<2
pg8000>=1.31,<2
